10-at-least-once-delivery/05-idempotent-event-handlers/main.go
package main

import "context"

type PaymentTaken struct {
	PaymentID string
	Amount    int
}

type PaymentsHandler struct {
	repo *PaymentsRepository
}

var processedPayments map[string]struct{}

func NewPaymentsHandler(repo *PaymentsRepository) *PaymentsHandler {
	return &PaymentsHandler{repo: repo}
}

func (p *PaymentsHandler) HandlePaymentTaken(ctx context.Context, event *PaymentTaken) error {
	return p.repo.SavePaymentTaken(ctx, event)
}

type PaymentsRepository struct {
	payments          []PaymentTaken
	processedPayments map[string]struct{}
}

func (p *PaymentsRepository) Payments() []PaymentTaken {
	return p.payments
}

func NewPaymentsRepository() *PaymentsRepository {
	return &PaymentsRepository{payments: []PaymentTaken{},
		processedPayments: make(map[string]struct{})}
}

func (p *PaymentsRepository) SavePaymentTaken(ctx context.Context, event *PaymentTaken) error {
	_, ok := p.processedPayments[event.PaymentID]
	if ok {
		return nil
	}
	p.processedPayments[event.PaymentID] = struct{}{}
	p.payments = append(p.payments, *event)
	return nil
}
